- Added missing initialization

+++ form-modeler-api/src/main/java/org/kie/formModeler/service/impl/FieldManagerImpl.java
@@ -47,7 +47,7 @@ public class FieldManagerImpl implements FieldManager {
     private Instance<FieldDefinition<Long>> longFields;
 
     @Inject
-    private Instance<FieldDefinition<Long>> shortFields;
+    private Instance<FieldDefinition<Short>> shortFields;
 
     @Inject
     private Instance<FieldDefinition<String>> stringFields;
